Change markdown files to create PDF using pandoc (#3)
* Change markdown files to create PDF using pandoc

Pandoc can now be used to create PDF out of the Markdown files

    pandoc --toc `cat toc.txt` -o inside-libreoffice.pdf

The toc.txt file should contain list of files in the correct order. A
rename that removes the space character in the file names can help to
simplify the pandoc invoke. This rename will be done later.

* Fix conflict resolution error

* Update author

Co-authored-by: Hossein <hossein@libreoffice.org>
Co-authored-by: flywire <flywire0@gmail.com>